Unused casts
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@1831630 13f79535-47bb-0310-9956-ffa450edef68

+++ batik-awt-util/src/main/java/org/apache/batik/ext/awt/geom/Polygon2D.java
@@ -293,7 +293,7 @@ public class Polygon2D implements Shape, Cloneable, Serializable {
      *                         <code>false</code> otherwise.
      */
     public boolean contains(int x, int y) {
-        return contains((double) x, (double) y);
+        return contains((double) x, y);
     }
 
     /**
